move InferVar into unify
